ControllerAdvice 완성  -안대식

+++ src/main/java/com/codeit/hrbank/global/exception/ErrorResponse.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 public record ErrorResponse(
 
         LocalDateTime timestamp,
-        Long status,
+        Integer status,
         String message,
         String details
